Handle help text printed by modulefiles
If the modulefiles at your site print a helpful message, e.g.
"Loading [module] with path [path]"
then module.load('module') will fail. This fixes that problem for me. I do not know if there are cases where things that are not os.environ calls might be part of the command. If that is the case, then a more elaborate fix would be needed.